Renames and starts generalizing conversation form
